requirements.txt
customtkinter
transformers
torch
accelerate
llama-cpp-python